--- src/components/Buttons/LapButton/LapButton.stories.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-
-import { LapButton, LapButtonProps } from './LapButton';
-
-export default {
-  component: LapButton,
-  title: 'Buttons/Lap',
-};
-
-const Template = (args: LapButtonProps) => <LapButton {...args} />;
-
-export const Default = Template.bind({}, { onClick() {} });
